# packages/contracts/internal/core-api.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/v1/routes/plans/{id}": {
      "get": {
        "tags": [
          "route-plan-controller"
        ],
        "operationId": "getOne",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePlanResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "route-plan-controller"
        ],
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoutePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePlanResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/routes/plans": {
      "get": {
        "tags": [
          "route-plan-controller"
        ],
        "operationId": "list_4",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseRoutePlanResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "route-plan-controller"
        ],
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoutePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePlanResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/routes/plans/{id}/cancel": {
      "post": {
        "tags": [
          "route-plan-controller"
        ],
        "operationId": "cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePlanResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/routes/plans/{id}/assign": {
      "post": {
        "tags": [
          "route-plan-controller"
        ],
        "operationId": "assign",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignDriverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePlanResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/routes/plans/suggest-order": {
      "post": {
        "tags": [
          "route-plan-controller"
        ],
        "operationId": "suggestOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestOrderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/routes/plans/stops/{stopId}/notify-passenger": {
      "post": {
        "tags": [
          "route-plan-controller"
        ],
        "operationId": "notifyPassenger",
        "parameters": [
          {
            "name": "stopId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/v1/routes/plans/stops/{stopId}/complete": {
      "post": {
        "tags": [
          "route-plan-controller"
        ],
        "operationId": "completeStop",
        "parameters": [
          {
            "name": "stopId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RouteStopResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/vehicles/{vehicleId}/cost-summary": {
      "get": {
        "tags": [
          "expense-entry-controller"
        ],
        "operationId": "summary",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/drivers/{driverId}/ratings/summary": {
      "get": {
        "tags": [
          "driver-rating-controller"
        ],
        "operationId": "summary_1",
        "parameters": [
          {
            "name": "driverId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DriverRatingSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StopInput": {
        "required": [
          "tipo"
        ],
        "type": "object",
        "properties": {
          "tipo": {
            "type": "string",
            "enum": [
              "COLETA",
              "ENTREGA"
            ]
          },
          "label": {
            "type": "string"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "collectionPointId": {
            "type": "string",
            "format": "uuid"
          },
          "janelaInicio": {
            "type": "string",
            "example": "08:00:00"
          },
          "janelaFim": {
            "type": "string",
            "example": "18:00:00"
          },
          "passengerId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "UpdateRoutePlanRequest": {
        "required": [
          "dataExecucao",
          "stops"
        ],
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "dataExecucao": {
            "type": "string",
            "format": "date"
          },
          "valor": {
            "type": "number"
          },
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StopInput"
            }
          }
        }
      },
      "RoutePlanResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "driverId": {
            "type": "string",
            "format": "uuid"
          },
          "driverName": {
            "type": "string"
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "vehiclePlate": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PLANEJADA",
              "EM_ANDAMENTO",
              "CONCLUIDA",
              "CANCELADA"
            ]
          },
          "categoria": {
            "type": "string",
            "enum": [
              "ROTA",
              "TRANSFER"
            ]
          },
          "dataExecucao": {
            "type": "string",
            "format": "date"
          },
          "valor": {
            "type": "number"
          },
          "custoEstimado": {
            "type": "number"
          },
          "margemRealizada": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RouteStopResponse"
            }
          },
          "viagemId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "RouteStopResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "COLETA",
              "ENTREGA"
            ]
          },
          "label": {
            "type": "string"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "collectionPointId": {
            "type": "string",
            "format": "uuid"
          },
          "janelaInicio": {
            "type": "string",
            "example": "08:00:00"
          },
          "janelaFim": {
            "type": "string",
            "example": "18:00:00"
          },
          "ordemSugerida": {
            "type": "integer",
            "format": "int32"
          },
          "ordemRealExecutada": {
            "type": "integer",
            "format": "int32"
          },
          "concluidaEm": {
            "type": "string",
            "format": "date-time"
          },
          "passengerId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CreateRoutePlanRequest": {
        "required": [
          "categoria",
          "dataExecucao",
          "stops"
        ],
        "type": "object",
        "properties": {
          "driverId": {
            "type": "string",
            "format": "uuid"
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "categoria": {
            "type": "string",
            "enum": [
              "ROTA",
              "TRANSFER"
            ]
          },
          "dataExecucao": {
            "type": "string",
            "format": "date"
          },
          "valor": {
            "type": "number"
          },
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StopInput"
            }
          },
          "viagemId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "AssignDriverRequest": {
        "required": [
          "driverId"
        ],
        "type": "object",
        "properties": {
          "driverId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "SuggestOrderRequest": {
        "required": [
          "stops"
        ],
        "type": "object",
        "properties": {
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StopInput"
            }
          }
        }
      },
      "SuggestOrderResponse": {
        "type": "object",
        "properties": {
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StopInput"
            }
          },
          "fallbackHaversine": {
            "type": "boolean"
          }
        }
      },
...
      "ExpenseSummaryResponse": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "totalValor": {
            "type": "number"
          },
          "odometerKm": {
            "type": "integer",
            "format": "int32"
          },
          "kmRodado": {
            "type": "integer",
            "format": "int32"
          },
          "custoPorKm": {
            "type": "number"
          }
        }
      },
...
      "PageResponseRoutePlanResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoutePlanResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "DriverRatingSummaryResponse": {
        "type": "object",
        "properties": {
          "driverId": {
            "type": "string",
            "format": "uuid"
          },
          "notaMedia": {
            "type": "number"
          },
          "totalAvaliacoes": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}